Create Zakaria_Ahmed_Student_Database.py
A python program that can be used to edit, as well as to access the information contained in a database stored in a file ( ex: .txt .dat) 
This database contains information about each student saved in a specific format in a line. A colon (:) separates the fields. For example, a student record with the following information:
ID: 1234
Name: Anne Shirley
Semester: 2
Year: 1
Number Of Courses: 5
Will be stored in this database as the following line:
1234:Anne Shirley:2:1:5